% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_compare_dict.R
\docType{data}
\name{id_var}
\alias{id_var}
\title{Comparison dictionaries and functions to build dictionaries}
\format{
An object of class \code{character} of length 1.
}
\usage{
id_var
}
\description{
Comparison dictionaries and functions to build dictionaries
}
\keyword{datasets}
