% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_comparison_set.R
\name{make_comparison_set}
\alias{make_comparison_set}
\title{Main function that takes an ID and builds out the comparison table}
\usage{
make_comparison_set(.data, .cols_comp, .labels)
}
\description{
Main function that takes an ID and builds out the comparison table
}
