% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_process.R
\name{redcap_process}
\alias{redcap_process}
\title{Like other reports, we begin with some modified calls to ADRCDash for processing purposes
Some functions for data processing}
\usage{
redcap_process()
}
\description{
Like other reports, we begin with some modified calls to ADRCDash for processing purposes
Some functions for data processing
}
