% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_process.R
\name{visit_read_in_alt}
\alias{visit_read_in_alt}
\title{Like other reports, we begin with some modified calls to ADRCDash for processing purposes
Some functions for data processing}
\usage{
visit_read_in_alt(
  token,
  synth = FALSE,
  dict = NULL,
  subtable_dict = NULL,
  use_redcap_factors = FALSE,
  ...
)
}
\description{
Like other reports, we begin with some modified calls to ADRCDash for processing purposes
Some functions for data processing
}
